+++ clarinet/services/image/__init__.py
@@ -8,8 +8,10 @@ from clarinet.services.image.grid_io import (
     classify_pair,
     read_grid,
 )
-from clarinet.services.image.image import FileType, Image, declare_nrrd_space
+from clarinet.services.image.image import FileType, Image
 from clarinet.services.image.layered_segmentation import LayeredSegmentation
+from clarinet.services.image.nrrd_repair import declare_nrrd_space
+from clarinet.services.image.nrrd_space import NrrdGrid, nrrd_grid_from_header
 from clarinet.services.image.orientation import OrientationUnverifiable, is_volume_misoriented
 from clarinet.services.image.segmentation import (
     PropName,
@@ -24,6 +26,7 @@ __all__ = [
     "GridRelation",
     "Image",
     "LayeredSegmentation",
+    "NrrdGrid",
     "OrientationUnverifiable",
     "PairVerdict",
     "PropName",
@@ -36,5 +39,6 @@ __all__ = [
     "declare_nrrd_space",
     "grid_relation",
     "is_volume_misoriented",
+    "nrrd_grid_from_header",
     "read_grid",
 ]